pyproject.toml
[tool.poetry]
name = "model-serving"
version = "0.1.0"
description = ""
authors = ["Dragonflip <jpmarquesdeoliveira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-stubs = "^4.2.6"
django-rest-framework = "^0.1.0"
markdown = "^3.5"
mlflow = "^2.8.0"
boto3 = "^1.28.71"
setuptools = "^68.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
